requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-jalali-date==1.0.1
djangorestframework==3.14.0
idna==3.4
jdatetime==4.1.0
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
redis==4.3.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
wrapt==1.14.1
